# backend/docs/swagger.yaml
components:
  schemas:
    api.HealthResponse:
      properties:
        status:
          example: healthy
          type: string
      type: object
    data:
      properties:
        data:
          $ref: '#/components/schemas/api.HealthResponse'
      type: object
...
    handler.QueryRequest:
      properties:
        connection_id:
          example: 9f1c7d2e4a6b4e89b88ad5f356bf7312
          type: string
        sql:
          example: SELECT 1
          type: string
        timeout_seconds:
          example: 30
          type: integer
      required:
      - connection_id
      - sql
      type: object
...
    response.APIError:
      properties:
        code:
          type: string
        message:
          type: string
        position:
          type: integer
      type: object
...
    response.ErrorEnvelope:
      properties:
        data:
          description: always null
        error:
          $ref: '#/components/schemas/response.APIError'
        meta:
          additionalProperties: {}
          type: object
        success:
          example: false
          type: boolean
      type: object
...
info:
  description: Local-first database GUI backend. All responses use the standard envelope
    (success, data, error, meta).
  title: DataDeck API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /query/execute:
    post:
      description: Runs a single SQL statement against a stored connection under a
        bounded timeout and records the attempt in history.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.QueryRequest'
              description: Query
              summary: request
        description: Query
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  data: {}
                  error:
                    $ref: '#/components/schemas/response.APIError'
                  meta:
                    additionalProperties: {}
                    type: object
                  success:
                    type: boolean
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Not Found
        "499":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Bad Gateway
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Gateway Timeout
      summary: Execute a SQL statement
      tags:
      - query
  /query/history:
    get:
      description: Returns recent executions newest-first, optionally filtered by
        connection.
      parameters:
      - description: Filter by connection id
        in: query
        name: connection_id
        schema:
          type: string
      - description: Page number (1-based)
        in: query
        name: page
        schema:
          type: integer
      - description: Page size (max 200)
        in: query
        name: page_size
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  data: {}
                  error:
                    $ref: '#/components/schemas/response.APIError'
                  meta:
                    additionalProperties: {}
                    type: object
                  success:
                    type: boolean
                type: object
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Internal Server Error
      summary: List query history
      tags:
      - query
servers:
- url: /api/v1